pyproject.toml
[project]
name = "e-masjid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "django>=4.2.20",
    "django-cors-headers>=4.4.0",
    "django-filter>=24.3",
    "django-redis>=5.4.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]
